requirements.txt
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.14.0
tqdm==4.66.1
pyyaml==6.0.1
